LPS-189055 Move and rename
